Configured based on apps

+++ apps/siege/ubuntu/stop.sh
@@ -2,13 +2,20 @@
 
 ###############################################################################
 # Environment data
+WEBTIER_SIEGE_WORKERS=${WEBTIER_SIEGE_WORKERS}
+WEBTIER_SIEGE_RUNMODE=${WEBTIER_SIEGE_RUNMODE}
+WEBTIER_SIEGE_WORDPRESS=${WEBTIER_SIEGE_WORDPRESS}
+WEBTIER_PATH=${WEBTIER_PATH}
+
 ###############################################################################
 
+. ${WEBTIER_PATH}/apps/common_func.sh
 
 ###############################################################################
 # Commands
 ###############################################################################
 
-
-killall run-siege
+if [ -z ${WEBTIER_SIEGE_WORDPRESS} ]; then
+    killall run-siege
+fi
 killall siege
